Move ContentManager utils to Utils.swift
